ready

+++ app/Http/Controllers/ArticleController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Article;
 use App\Category;
+use Hekmatinasser\Verta\Verta;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 class ArticleController extends Controller
@@ -23,12 +25,22 @@ class ArticleController extends Controller
     }
     function delete($id)
     {
-    $article = Article::FindOrFail($id);
-    $article->delete();
-    return redirect(route('home'));
+        if (Auth::user()->id != 1){
+            abort(403);
+        }
+        $article = Article::FindOrFail($id);
+        if($article->file!=null) {
+            $file_path = public_path('file/') . $article->id . '.' . $article->file;
+            @unlink($file_path);
+        }
+        $article->delete();
+        return redirect(route('home'));
     }
     public function clear_hits($id)
     {
+        if (Auth::user()->id != 1){
+            abort(403);
+        }
         $article = Article::FindOrFail($id);
         $article->hits =100;
         $article->save();
@@ -37,5 +49,38 @@ class ArticleController extends Controller
     public function form()
     {
         return view('admin.article');
+    }
+    public function create(Request $request)
+    {
+        $this->Validate($request, [
+            'category' => 'required|integer',
+            'title' => 'required|string',
+            'summary' => 'required|string',
+            'provider' => 'required|string',
+        ]);
+
+        if($request->hasFile('file')){
+            $file = $request->file('file')->getClientOriginalExtension();
+        }else{
+            $file = null;
+        }
+
+        $article = new Article(array(
+            'category_id' => $request->category,
+            'title' => $request->title,
+            'summary' => $request->summary,
+            'provider' => $request->provider,
+            'date_time' => Verta::now(),
+            'hits' => 100,
+            'file' => $file ,
+        ));
+        $article->save();
+
+        $file_name = $article->id .'.'.$file;
+        if($request->hasFile('file')) {
+            $request->file('file')->move(public_path('file/'), $file_name);
+        }
+        return redirect(route('home'));
+
     }
 }
