+++ app/Http/Controllers/CategoryController.php
@@ -5,12 +5,13 @@ namespace App\Http\Controllers;
 use App\Article;
 use App\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
     public function index($id)
     {
-        $articles = Article::where('category_id','=',$id)->get();
+        $articles = Article::where('category_id','=',$id)->orderby('date_time','desc')->get();
         $category = Category::FindOrFail($id);
         $category->hits ++;
         $category->save();
@@ -18,12 +19,26 @@ class CategoryController extends Controller
     }
     function delete($id)
     {
+        if (Auth::user()->id != 1){
+            abort(403);
+        }
         $category = Category::FindOrFail($id);
+        $articles = Article::where('category_id','=',$id)->get();
+        foreach ($articles as $article){
+            if($article->file!=null) {
+                $file_path = public_path('file/') . $article->id . '.' . $article->file;
+                @unlink($file_path);
+            }
+            $article->delete();
+        }
         $category->delete();
         return redirect(route('home'));
     }
     public function clear_hits($id)
     {
+        if (Auth::user()->id != 1){
+            abort(403);
+        }
         $category = Category::FindOrFail($id);
         $category->hits =100;
         $category->save();
